Abstract browser stuff to core module to allow reuse in other modules
